Fix part of #659
There are still issues here, but I think it's probably in JS: when focus is lost on the textfield, the `mat-floating-label--float-above` is removed, irrespective of existance of text or placeholder. Needs some more investigation.

What's fixed here is the situation where overlap occurs when the value is not just whitespace or when the placeholder is not just whitespace and should be visible.