#713 - Task Canceled Exception
This is an issue with JSInterop, where it natively awaits the JS task, but does not throw an exception to show that the task timed out. 
https://github.com/dotnet/aspnetcore/issues/18875
https://github.com/dotnet/aspnetcore/issues/21384
The second issue in the description indicates that MS will need to do something to allow a more specific exception that can be caught in the try/catch, though the TaskCanceledException is enough for now.